fix: allow Python version to be passed via environment or config file
The -p/--python option of virtualenv's builtin discovery plugin can be
omitted, in which case it will use the `python` setting in the config
file or the `VIRTUALENV_PYTHON` environment variable. However, the same
is not true of virtualenv-pyenv's discovery plugin. In this commit I'm
copying the call to parser.add_argument() from the builtin discovery
plugin, so that virtualenv-pyenv's -p/--python option will have the same
semantics as the builtin one: its default value will be taken from
the config file or environment if the option is not given on the command
line. This allows virtualenv-pyenv to work with other projects like tox
which use the environment variable to communicate which Python version
to use.

As a complement to that, I added logic to raise an error if no Python
version is available at all.

Note that the logic to pull default values from the config file or
environment is implemented in virtualenv itself, but it only triggers if
the action registered to handle the action has a default value which is
not argparse.SUPPRESS. (Technically it also requires the action to have
a destination, but pretty much all actions do.) So the minimal change
that would make this feature work is just adding a default value to
the call to parser.add_argument(). But the default value is pretty much
the only meaningful difference between the add_argument() calls in
the builtin discovery plugin and this one, so I thought it would make
sense to just copy the entire call for consistency.